Separate dormyboba service from server
