+++ dormyboba_core/service/__init__.py
@@ -1,0 +1,1 @@
+from .dormyboba_servicer import *
